Finished commit

+++ src/main/java/pro/sky/java/course2/calculatornew/controller/service/ServiceCalculator.java
@@ -3,9 +3,9 @@ package pro.sky.java.course2.calculatornew.controller.service;
 public interface ServiceCalculator {
     String hello();
 
-    String plus(int num1, int num2);
-    String minus(int num1, int num2);
-    String multiply(int num1, int num2);
-    String divide(int num1, int num2);
+    String plus(Double num1, Double num2);
+    String minus(Double num1, Double num2);
+    String multiply(Double num1, Double num2);
+    String divide(Double num1, Double num2);
 
 }
